# shared/openapi/amon-v1-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amon API",
    "description": "Amon v1 privacy-first retrieval backend. Queries and page content are handled transiently.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search",
        "operationId": "search_v1_search_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/research": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Research",
        "operationId": "research_v1_research_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemSourcePayload": {
        "properties": {
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "page_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Title"
          },
          "cleaned_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cleaned Excerpt"
          },
          "bullet_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bullet Points"
          },
          "typed_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typed Metadata"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url",
          "domain"
        ],
        "title": "ItemSourcePayload"
      },
      "ModelInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "name",
          "version"
        ],
        "title": "ModelInfo"
      },
      "ProviderInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "provider_result_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Result Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProviderInfo"
      },
      "ResearchRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "prompt_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Context"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemSourcePayload"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 2,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "title",
          "items"
        ],
        "title": "ResearchRequest"
      },
      "ResearchResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary_text": {
            "type": "string",
            "title": "Summary Text"
          },
          "bullet_summary": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bullet Summary"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ResearchSourceRef"
            },
            "type": "array",
            "title": "Sources"
          },
          "model": {
            "$ref": "#/components/schemas/ModelInfo"
          }
        },
        "type": "object",
        "required": [
          "title",
          "summary_text",
          "bullet_summary",
          "sources",
          "model"
        ],
        "title": "ResearchResponse"
      },
      "ResearchSourceRef": {
        "properties": {
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          }
        },
        "type": "object",
        "title": "ResearchSourceRef"
      },
...
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Query"
          },
          "count": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Count",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "SearchResponse"
      },
      "SearchResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "result_type": {
            "type": "string",
            "enum": [
              "web_page",
              "article",
              "place",
              "product"
            ],
            "title": "Result Type",
            "default": "web_page"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "typed_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typed Metadata"
          },
          "provider": {
            "$ref": "#/components/schemas/ProviderInfo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "domain",
          "provider"
        ],
        "title": "SearchResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}